Move ChangeSet to its own module
Just moving the code + a few new public methods to avoid breaking
encapsulation.
