+++ icechunk/src/lib.rs
@@ -17,6 +17,7 @@
 //!     - a caching wrapper implementation
 //! - The datastructures are represented by concrete types in the [`mod@format`] modules.
 //!   These datastructures use Arrow RecordBatches for representation.
+pub mod change_set;
 pub mod format;
 pub mod metadata;
 pub mod refs;
